feat: Update Config component schema to be more flexible
